+++ package/src/jder_fastapi/responses/json/__init__.py
@@ -4,14 +4,16 @@ from .main import (
     CreateJsonSuccessResponseOptions,
     JsonResponse,
     JsonResponseError,
+    JsonResponseErrorInput,
     createJsonResponse,
 )
 
 __all__ = [
-    "JsonResponseError",
-    "JsonResponse",
+    "CreateJsonFailureResponseOptions",
     "CreateJsonResponseBaseOptions",
     "CreateJsonSuccessResponseOptions",
-    "CreateJsonFailureResponseOptions",
+    "JsonResponse",
+    "JsonResponseError",
+    "JsonResponseErrorInput",
     "createJsonResponse",
 ]